+++ src/competitive_verifier/oj/languages/cplusplus.py
