+++ src/competitive_verifier/oj/languages/go.py
