+++ src/competitive_verifier/oj/languages/nim.py
