--- src/competitive_verifier/oj/verify/__init__.py
@@ -1,1 +1,0 @@
-"""Port of oj-verify."""
